change types

+++ src/main/resources/db/migration/V1_0__initial_structure.sql
@@ -51,6 +51,6 @@ CREATE TABLE transactions
     to_address         VARCHAR(64) NOT NULL,
     block_number       BIGINT      NOT NULL,
     date               TIMESTAMP   NOT NULL,
-    blockchain_type_id BIGINT REFERENCES blockchain_types,
+    blockchain_type_id INT REFERENCES blockchain_types,
     UNIQUE (hash, blockchain_type_id)
 );
